chore: move files from ./src to ./examples
